adjust hamburger_menu_option helper function to act as a helper, not as test. Moved the test part into the test case
