+++ iris/api/helpers/images/linux/hamburger_menu.png
